refactor:  모바일 페이지 컴포넌트 분리 (#476)
