+++ frontend/src/components/LoginPageUI/LoginPageUI.tsx
@@ -1,7 +1,7 @@
 import styled, { css } from 'styled-components';
 
 import Footer from '~/components/@common/Footer';
-import Header from '~/components/@common/Header';
+import { Header } from '~/components/@common/Header';
 
 import useMediaQuery from '~/hooks/useMediaQuery';
